// web/admin/src/generated-client/model/getLatestEntriesRow.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 1.0.0
 */

export interface GetLatestEntriesRow {
  Path: string;
  Title: string;
  Body: string;
  Visibility: string;
  Format: string;
  /** @nullable */
  PublishedAt?: string | null;
  /** @nullable */
  LastEditedAt?: string | null;
  /** @nullable */
  CreatedAt?: string | null;
  /** @nullable */
  UpdatedAt?: string | null;
  /** @nullable */
  ImageUrl?: string | null;
}